move streamwalker to separate file
