+++ src/rsrc_mgr/Makefile.am
@@ -83,6 +83,7 @@ marfs_rman_SOURCES =        \
     resourceprocessing.c    \
     resourcethreads.c       \
     rmanstate.c             \
+    streamwalker.c          \
     summary_log_setup.c     \
     work.c                  \
     worker.c
@@ -97,11 +98,11 @@ test_resourcelog_SOURCES = testing/test_resourcelog.c
 test_resourcelog_LDADD = libResourceLog.la ../logging/liblogging.la
 test_resourcelog_CFLAGS = $(XML_CFLAGS)
 
-test_resourceprocessing_SOURCES = testing/test_resourceprocessing.c repack.c
+test_resourceprocessing_SOURCES = testing/test_resourceprocessing.c repack.c streamwalker.c
 test_resourceprocessing_LDADD = libResourceLog.la ../datastream/libDatastream.la ../logging/liblogging.la
 test_resourceprocessing_CFLAGS = $(XML_CFLAGS)
 
-test_resourcethreads_SOURCES = testing/test_resourcethreads.c repack.c resourceprocessing.c
+test_resourcethreads_SOURCES = testing/test_resourcethreads.c repack.c resourceprocessing.c streamwalker.c
 test_resourcethreads_LDADD = libResourceLog.la ../datastream/libDatastream.la ../logging/liblogging.la
 test_resourcethreads_CFLAGS = $(XML_CFLAGS)
