+++ src/rsrc_mgr/resourceprocessing.h
@@ -63,45 +63,7 @@ GNU licenses can be found at http://www.gnu.org/licenses/.
 
 #include "datastream/datastream.h"
 #include "rsrc_mgr/repack.h"
-
-typedef struct {
-   time_t gcthreshold;      // files newer than this will not be GCd
-                            //    Recommendation -- this should be fairly old ( # of days ago )
-   time_t repackthreshold;  // files newer than this will not be repacked
-                            //    Recommendation -- this should be quite recent ( # of minutes ago )
-   time_t rebuildthreshold; // files newer than this will not be rebuilt ( have data errors repaired )
-                            //    Recommendation -- this should be quite recent ( # of minutes ago )
-   time_t cleanupthreshold; // files newer than this will not be cleaned up ( i.e. repack marker files )
-                            //    Recommendation -- this should be semi-recent ( # of hours ago )
-   // NOTE -- setting any of these values too close to the current time risks undefined resource
-   //         processing behavior ( i.e. trying to cleanup ongoing repacks )
-   // NOTE -- setting any of these values to zero will cause the corresponding operations to be skipped
-} thresholds;
-
-typedef struct {
-   // quota info
-   size_t fileusage;   // count of active files
-   size_t byteusage;   // count of active bytes
-   // stream info
-   size_t filecount;   // count of files in the datastream
-   size_t objcount;    // count of objects in the datastream
-   size_t bytecount;   // count of all bytes in the datastream
-   size_t streamcount; // count of all streams ( always one, here, but having this value in the struct helps elsewhere )
-   // GC info
-   size_t delobjs;    // count of deleted objects
-   size_t delfiles;   // count of deleted files
-   size_t delstreams; // count of completely deleted datastreams
-   size_t volfiles;   // count of 'volatile' files ( those deleted too recently for gc )
-   // repack info
-   size_t rpckfiles;  // count of files repacked
-   size_t rpckbytes;  // count of bytes repacked
-   // rebuild info
-   size_t rbldobjs;   // count of rebuilt objects
-   size_t rbldbytes;  // count of rebuilt bytes
-} streamwalker_report;
-
-// forward decls of internal types
-typedef struct streamwalker* streamwalker;
+#include "rsrc_mgr/streamwalker.h"
 
 //   -------------   RESOURCE PROCESSING FUNCTIONS    -------------
 
@@ -139,43 +101,6 @@ int process_refdir( marfs_ns* ns, MDAL_SCANNER refdir, const char* refdirpath, c
  * @return opinfo* : Reference to the newly generated op, or NULL on failure
  */
 opinfo* process_rebuildmarker( marfs_position* pos, char* markerpath, time_t rebuildthresh, size_t objno );
-
-/**
- * Open a streamwalker based on the given fileno zero reference target
- * @param streamwalker* swalker : Reference to be populated with the produced streamwalker
- * @param marfs_position* pos : MarFS position to be used by this walker
- * @param const char* reftgt : Reference path of the first ( fileno zero ) file of the datastream
- * @param thresholds thresh : Threshold values to be used for determining operation targets
- * @param ne_location* rebuildloc : Location-based rebuild target
- * @return int : Zero on success, or -1 on failure
- *               NOTE -- It is possible for success to be indicated without any streamwalker being produced,
- *                       such as, if the stream is incomplete ( missing FTAG values ).
- *                       Failure will only be indicated if an unexpected condition occurred, such as, if the
- *                       datastream is improperly formatted.
- */
-int process_openstreamwalker( streamwalker* swalker, marfs_position* pos, const char* reftgt, thresholds thresh, ne_location* rebuildloc );
-
-/**
- * Iterate over a datastream, accumulating quota values and identifying operation targets
- * NOTE -- This func will return all possible operations, given walker settings.  It is up to the caller whether those ops
- *         will actually be executed via process_operation().
- * @param streamwalker* swalker : Reference to the streamwalker to be iterated
- * @param opinfo** gcops : Reference to be populated with generated GC operations
- * @param opinfo** repackops : Reference to be populated with generated repack operations
- * @param opinfo** rebuildops : Reference to be populated with generated rebuild operations
- * @return int : 0, if the end of the datastream was reached and no new operations were generated;
- *               1, if new operations were generated by this iteration;
- *               -1, if a failure occurred
- */
-int process_iteratestreamwalker( streamwalker* swalker, opinfo** gcops, opinfo** repackops, opinfo** rebuildops );
-
-/**
- * Close the given streamwalker
- * @param streamwalker* swalker : Reference to the streamwalker to be closed
- * @param streamwalker_report* report : Reference to a report to be populated with final counts
- * @return int : Zero on success, 1 if the walker was closed prior to iteration completion, or -1 on failure
- */
-int process_closestreamwalker( streamwalker* swalker, streamwalker_report* report );
 
 /**
  * Perform the given operation
