+++ src/rsrc_mgr/testing/test_resourceprocessing.c
@@ -384,7 +384,7 @@ int main(int argc, char **argv)
    };
 
    streamwalker walker = NULL;
-   if (process_openstreamwalker(&walker, &pos, rpath, thresh, &rebuildloc)) {
+   if (streamwalker_open(&walker, &pos, rpath, thresh, &rebuildloc)) {
       printf("failed to open streamwalker for \"%s\"\n", rpath);
       return -1;
    }
@@ -394,7 +394,7 @@ int main(int argc, char **argv)
    opinfo* rebuildops = NULL;
    ssize_t rebuildcount = 0;
    int iterres = 1;
-   while((iterres = process_iteratestreamwalker(&walker, &gcops, &repackops, &rebuildops)) > 0) {
+   while((iterres = streamwalker_iterate(&walker, &gcops, &repackops, &rebuildops)) > 0) {
       // not expecting any gc or repack ops, at present
       if (gcops) {
          printf("unexpected GCOPS following first traversal of no-pack stream\n");
@@ -432,7 +432,7 @@ int main(int argc, char **argv)
    }
 
    streamwalker_report walkreport;
-   if (process_closestreamwalker(&walker, &walkreport)) {
+   if (streamwalker_close(&walker, &walkreport)) {
       printf("failed to close first streamwalker\n");
       return -1;
    }
@@ -507,7 +507,7 @@ int main(int argc, char **argv)
    thresh.repackthreshold = 0;
    thresh.rebuildthreshold = 0;
    thresh.cleanupthreshold = 0;
-   if (process_openstreamwalker(&walker, &pos, rpath, thresh, NULL)) {
+   if (streamwalker_open(&walker, &pos, rpath, thresh, NULL)) {
       printf("failed to open streamwalker3 for \"%s\"\n", rpath);
       return -1;
    }
@@ -515,7 +515,7 @@ int main(int argc, char **argv)
    gcops = NULL;
    repackops = NULL;
    rebuildops = NULL;
-   if (process_iteratestreamwalker(&walker, &gcops, &repackops, &rebuildops)) {
+   if (streamwalker_iterate(&walker, &gcops, &repackops, &rebuildops)) {
       printf("unexpected result of first quota-only iteration from \"%s\"\n", rpath);
       return -1;
    }
@@ -536,7 +536,7 @@ int main(int argc, char **argv)
       return -1;
    }
 
-   if (process_closestreamwalker(&walker, &walkreport)) {
+   if (streamwalker_close(&walker, &walkreport)) {
       printf("failed to close walker after second gc iteration\n");
       return -1;
    }
@@ -567,7 +567,7 @@ int main(int argc, char **argv)
    thresh.repackthreshold = currenttime.tv_sec + 120;
    thresh.rebuildthreshold = 0; // no rebuilds for now
    thresh.cleanupthreshold = currenttime.tv_sec + 120;
-   if (process_openstreamwalker(&walker, &pos, rpath, thresh, NULL)) {
+   if (streamwalker_open(&walker, &pos, rpath, thresh, NULL)) {
       printf("failed to open streamwalker2 for \"%s\"\n", rpath);
       return -1;
    }
@@ -575,7 +575,7 @@ int main(int argc, char **argv)
    gcops = NULL;
    repackops = NULL;
    rebuildops = NULL;
-   if (process_iteratestreamwalker(&walker, &gcops, &repackops, &rebuildops) < 1) {
+   if (streamwalker_iterate(&walker, &gcops, &repackops, &rebuildops) < 1) {
       printf("unexpected result of second iteration from \"%s\"\n", rpath);
       return -1;
    }
@@ -610,12 +610,12 @@ int main(int argc, char **argv)
    }
 
    // continue iteration, which should produce no further ops
-   if (process_iteratestreamwalker(&walker, &gcops, &repackops, &rebuildops)) {
+   if (streamwalker_iterate(&walker, &gcops, &repackops, &rebuildops)) {
       printf("unexpected walker iteration after first gc from nopack\n");
       return -1;
    }
 
-   if (process_closestreamwalker(&walker, NULL)) {
+   if (streamwalker_close(&walker, NULL)) {
       printf("failed to close walker after second gc iteration\n");
       return -1;
    }
@@ -708,7 +708,7 @@ int main(int argc, char **argv)
    thresh.repackthreshold = currenttime.tv_sec + 120;
    thresh.rebuildthreshold = 0; // no rebuilds for now
    thresh.cleanupthreshold = currenttime.tv_sec + 120;
-   if (process_openstreamwalker(&walker, &pos, rpath, thresh, NULL)) {
+   if (streamwalker_open(&walker, &pos, rpath, thresh, NULL)) {
       printf("failed to open streamwalker3 for \"%s\"\n", rpath);
       return -1;
    }
@@ -716,7 +716,7 @@ int main(int argc, char **argv)
    gcops = NULL;
    repackops = NULL;
    rebuildops = NULL;
-   if (process_iteratestreamwalker(&walker, &gcops, &repackops, &rebuildops) < 1) {
+   if (streamwalker_iterate(&walker, &gcops, &repackops, &rebuildops) < 1) {
       printf("unexpected result of third iteration from \"%s\"\n", rpath);
       return -1;
    }
@@ -738,7 +738,7 @@ int main(int argc, char **argv)
    }
 
    // continue iteration, which should produce no further ops
-   if (process_iteratestreamwalker(&walker, &gcops, &repackops, &rebuildops)) {
+   if (streamwalker_iterate(&walker, &gcops, &repackops, &rebuildops)) {
       printf("unexpected walker iteration after second gc from nopack\n");
       return -1;
    }
@@ -756,7 +756,7 @@ int main(int argc, char **argv)
       return -1;
    }
 
-   if (process_closestreamwalker(&walker, NULL)) {
+   if (streamwalker_close(&walker, NULL)) {
       printf("failed to close walker after second gc iteration\n");
       return -1;
    }
@@ -800,7 +800,7 @@ int main(int argc, char **argv)
    thresh.repackthreshold = currenttime.tv_sec + 120;
    thresh.rebuildthreshold = 0; // no rebuilds for now
    thresh.cleanupthreshold = currenttime.tv_sec + 120;
-   if (process_openstreamwalker(&walker, &pos, rpath, thresh, NULL)) {
+   if (streamwalker_open(&walker, &pos, rpath, thresh, NULL)) {
       printf("failed to open streamwalker3 for \"%s\"\n", rpath);
       return -1;
    }
@@ -808,7 +808,7 @@ int main(int argc, char **argv)
    gcops = NULL;
    repackops = NULL;
    rebuildops = NULL;
-   if (process_iteratestreamwalker(&walker, &gcops, &repackops, &rebuildops) < 1) {
+   if (streamwalker_iterate(&walker, &gcops, &repackops, &rebuildops) < 1) {
       printf("unexpected result of third iteration from \"%s\"\n", rpath);
       return -1;
    }
@@ -830,12 +830,12 @@ int main(int argc, char **argv)
    }
 
    // continue iteration, which should produce no further ops
-   if (process_iteratestreamwalker(&walker, &gcops, &repackops, &rebuildops)) {
+   if (streamwalker_iterate(&walker, &gcops, &repackops, &rebuildops)) {
       printf("unexpected walker iteration after third gc from nopack\n");
       return -1;
    }
 
-   if (process_closestreamwalker(&walker, NULL)) {
+   if (streamwalker_close(&walker, NULL)) {
       printf("failed to close walker after second gc iteration\n");
       return -1;
    }
@@ -1118,7 +1118,7 @@ int main(int argc, char **argv)
    thresh.repackthreshold = 0;
    thresh.rebuildthreshold = 0;
    thresh.cleanupthreshold = 0;
-   if (process_openstreamwalker(&walker, &pos, rpath, thresh, NULL)) {
+   if (streamwalker_open(&walker, &pos, rpath, thresh, NULL)) {
       printf("failed to open quick-streamwalker1 for \"%s\"\n", rpath);
       return -1;
    }
@@ -1126,7 +1126,7 @@ int main(int argc, char **argv)
    gcops = NULL;
    repackops = NULL;
    rebuildops = NULL;
-   if (process_iteratestreamwalker(&walker, &gcops, &repackops, &rebuildops)) {
+   if (streamwalker_iterate(&walker, &gcops, &repackops, &rebuildops)) {
       printf("unexpected result of first quota-only iteration from \"%s\"\n", rpath);
       return -1;
    }
@@ -1147,7 +1147,7 @@ int main(int argc, char **argv)
       return -1;
    }
 
-   if (process_closestreamwalker(&walker, &walkreport)) {
+   if (streamwalker_close(&walker, &walkreport)) {
       printf("failed to close walker after first quick iteration of pack\n");
       return -1;
    }
@@ -1173,7 +1173,7 @@ int main(int argc, char **argv)
    thresh.repackthreshold = currenttime.tv_sec + 120;
    thresh.rebuildthreshold = 0; // no rebuilds for now
    thresh.cleanupthreshold = currenttime.tv_sec + 120;
-   if (process_openstreamwalker(&walker, &pos, rpath, thresh, NULL)) {
+   if (streamwalker_open(&walker, &pos, rpath, thresh, NULL)) {
       printf("failed to open streamwalker2 for \"%s\"\n", rpath);
       return -1;
    }
@@ -1181,7 +1181,7 @@ int main(int argc, char **argv)
    gcops = NULL;
    repackops = NULL;
    rebuildops = NULL;
-   if (process_iteratestreamwalker(&walker, &gcops, &repackops, &rebuildops)) {
+   if (streamwalker_iterate(&walker, &gcops, &repackops, &rebuildops)) {
       printf("unexpected result of second iteration from \"%s\"\n", rpath);
       return -1;
    }
@@ -1202,7 +1202,7 @@ int main(int argc, char **argv)
       return -1;
    }
 
-   if (process_closestreamwalker(&walker, &walkreport)) {
+   if (streamwalker_close(&walker, &walkreport)) {
       printf("failed to close walker after second gc iteration\n");
       return -1;
    }
@@ -1340,7 +1340,7 @@ int main(int argc, char **argv)
    thresh.repackthreshold = currenttime.tv_sec + 120;
    thresh.rebuildthreshold = 0; // no rebuilds for now
    thresh.cleanupthreshold = currenttime.tv_sec + 120;
-   if (process_openstreamwalker(&walker, &pos, rpath, thresh, NULL)) {
+   if (streamwalker_open(&walker, &pos, rpath, thresh, NULL)) {
       printf("failed to open streamwalker3 for \"%s\"\n", rpath);
       return -1;
    }
@@ -1348,7 +1348,7 @@ int main(int argc, char **argv)
    gcops = NULL;
    repackops = NULL;
    rebuildops = NULL;
-   if (process_iteratestreamwalker(&walker, &gcops, &repackops, &rebuildops) < 1) {
+   if (streamwalker_iterate(&walker, &gcops, &repackops, &rebuildops) < 1) {
       printf("unexpected result of third iteration from \"%s\"\n", rpath);
       return -1;
    }
@@ -1370,11 +1370,11 @@ int main(int argc, char **argv)
    }
 
    // continue iteration, which should produce no further ops
-   if (process_iteratestreamwalker(&walker, &gcops, &repackops, &rebuildops)) {
+   if (streamwalker_iterate(&walker, &gcops, &repackops, &rebuildops)) {
       printf("unexpected walker iteration after third gc from pack\n");
       return -1;
    }
-   if (process_closestreamwalker(&walker, &walkreport)) {
+   if (streamwalker_close(&walker, &walkreport)) {
       printf("failed to close walker after third pack gc iteration\n");
       return -1;
    }
@@ -1469,7 +1469,7 @@ int main(int argc, char **argv)
    free(objname2);
 
    // walk a final time, to delete the entire stream
-   if (process_openstreamwalker(&walker, &pos, rpath, thresh, NULL)) {
+   if (streamwalker_open(&walker, &pos, rpath, thresh, NULL)) {
       printf("failed to open streamwalker4 for \"%s\"\n", rpath);
       return -1;
    }
@@ -1477,7 +1477,7 @@ int main(int argc, char **argv)
    gcops = NULL;
    repackops = NULL;
    rebuildops = NULL;
-   if (process_iteratestreamwalker(&walker, &gcops, &repackops, &rebuildops) < 1) {
+   if (streamwalker_iterate(&walker, &gcops, &repackops, &rebuildops) < 1) {
       printf("unexpected result of fourth iteration from \"%s\"\n", rpath);
       return -1;
    }
@@ -1499,11 +1499,11 @@ int main(int argc, char **argv)
    }
 
    // continue iteration, which should produce no further ops
-   if (process_iteratestreamwalker(&walker, &gcops, &repackops, &rebuildops)) {
+   if (streamwalker_iterate(&walker, &gcops, &repackops, &rebuildops)) {
       printf("unexpected walker iteration after fourth gc from pack\n");
       return -1;
    }
-   if (process_closestreamwalker(&walker, &walkreport)) {
+   if (streamwalker_close(&walker, &walkreport)) {
       printf("failed to close walker after fourth pack gc iteration\n");
       return -1;
    }
@@ -1805,7 +1805,7 @@ int main(int argc, char **argv)
    thresh.repackthreshold = currenttime.tv_sec + 120;
    thresh.rebuildthreshold = 0; // no rebuilds for now
    thresh.cleanupthreshold = currenttime.tv_sec + 120;
-   if (process_openstreamwalker(&walker, &pos, rpath, thresh, NULL)) {
+   if (streamwalker_open(&walker, &pos, rpath, thresh, NULL)) {
       printf("failed to open streamwalker1 for \"%s\"\n", rpath);
       return -1;
    }
@@ -1813,7 +1813,7 @@ int main(int argc, char **argv)
    gcops = NULL;
    repackops = NULL;
    rebuildops = NULL;
-   if (process_iteratestreamwalker(&walker, &gcops, &repackops, &rebuildops) < 1) {
+   if (streamwalker_iterate(&walker, &gcops, &repackops, &rebuildops) < 1) {
       printf("unexpected result of fifth iteration from \"%s\"\n", rpath);
       return -1;
    }
@@ -1835,7 +1835,7 @@ int main(int argc, char **argv)
    }
 
    // continue iteration, which should produce no further ops
-   if (process_iteratestreamwalker(&walker, &gcops, &repackops, &rebuildops)) {
+   if (streamwalker_iterate(&walker, &gcops, &repackops, &rebuildops)) {
       printf("unexpected walker iteration after first gc from parallel-write\n");
       return -1;
    }
@@ -1853,7 +1853,7 @@ int main(int argc, char **argv)
       return -1;
    }
 
-   if (process_closestreamwalker(&walker, &walkreport)) {
+   if (streamwalker_close(&walker, &walkreport)) {
       printf("failed to close walker after first parallel-write gc iteration\n");
       return -1;
    }
@@ -1964,7 +1964,7 @@ int main(int argc, char **argv)
    thresh.repackthreshold = currenttime.tv_sec + 120;
    thresh.rebuildthreshold = 0; // no rebuilds for now
    thresh.cleanupthreshold = currenttime.tv_sec + 120;
-   if (process_openstreamwalker(&walker, &pos, rpath, thresh, NULL)) {
+   if (streamwalker_open(&walker, &pos, rpath, thresh, NULL)) {
       printf("failed to open streamwalker1 for \"%s\"\n", rpath);
       return -1;
    }
@@ -1972,7 +1972,7 @@ int main(int argc, char **argv)
    gcops = NULL;
    repackops = NULL;
    rebuildops = NULL;
-   if (process_iteratestreamwalker(&walker, &gcops, &repackops, &rebuildops) < 1) {
+   if (streamwalker_iterate(&walker, &gcops, &repackops, &rebuildops) < 1) {
       printf("unexpected result of final iteration from \"%s\"\n", rpath);
       return -1;
    }
@@ -1994,7 +1994,7 @@ int main(int argc, char **argv)
    }
 
    // continue iteration, which should produce no further ops
-   if (process_iteratestreamwalker(&walker, &gcops, &repackops, &rebuildops)) {
+   if (streamwalker_iterate(&walker, &gcops, &repackops, &rebuildops)) {
       printf("unexpected walker iteration after final gc from parallel-write\n");
       return -1;
    }
@@ -2012,7 +2012,7 @@ int main(int argc, char **argv)
       return -1;
    }
 
-   if (process_closestreamwalker(&walker, &walkreport)) {
+   if (streamwalker_close(&walker, &walkreport)) {
       printf("failed to close walker after final parallel-write gc iteration\n");
       return -1;
    }
